requirements/web.txt
-r base.txt

# web-facing libraries (Not used for lambda clients)
django-allauth==0.26.1
django-loginas==0.2.2
django-webpack-loader==0.3.3
git+https://github.com/jleclanche/django-cloud-browser.git#egg=django_cloud_browser
humanize==0.5.1
libsass==0.11.1
nodeenv==0.13.6

# redis requirements
django-rq==0.9.2
django-rq-dashboard==0.3.0
redis==2.10.5
rq==0.6.0
